+++ Sources/Paywall/Paywall/Events/Internal Tracking/Trackable Events/SuperwallEvent.swift
@@ -196,7 +196,8 @@ enum SuperwallEvent {
           "result": "present",
           "trigger_name": triggerName
         ]
-      case .unknownEvent:
+      case .triggerNotFound,
+        .error:
         return [:]
       }
     }