+++ CHANGELOG.md
@@ -6,9 +6,11 @@ The changelog for `Paywall`. Also see the [releases](https://github.com/superwal
 
 ### Breaking Changes
 
-- Renamed `Paywall.trigger` to `Paywall.track`. We found that having separate implicit (`Paywall.track(...)`) and explicit (`Paywall.trigger(...)`) trigger functions caused confusion. So from now on, you'll just use `Paywall.track` for all events within your app.
+- Renamed `Paywall.trigger` to `Paywall.track`. We found that having separate implicit (`Paywall.track(...)`) and explicit (`Paywall.trigger(...)`) trigger functions caused confusion. So from now on, you'll just use `Superwall.track` for all events within your app.
 
 ### Enhancements
+
+- New function `Superwall.track(event:params:overrides)` which returns an `AnyPublisher<PaywallState, Never>` for those Combine lovers. We've updated our sample apps to show you how to use that.
 
 ### Fixes